requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
databases
bcrypt
python-dotenv
aiomysql
